+++ libp2p/src/builder/phase/quic.rs
@@ -171,18 +171,6 @@ impl<T: AuthenticatedMultiplexedTransport> SwarmBuilder<super::provider::Tokio, 
         self,
         cfg: libp2p_dns::ResolverConfig,
         opts: libp2p_dns::ResolverOpts,
-    ) -> SwarmBuilder<super::provider::Tokio, WebsocketPhase<impl AuthenticatedMultiplexedTransport>>
-    {
-        self.without_quic()
-            .without_any_other_transports()
-            .with_dns_config(cfg, opts)
-    }
-
-    /// See [`SwarmBuilder::try_with_dns_config`].
-    pub fn try_with_dns_config(
-        self,
-        cfg: libp2p_dns::ResolverConfig,
-        opts: libp2p_dns::ResolverOpts,
     ) -> Result<
         SwarmBuilder<
             super::provider::Tokio,
@@ -192,7 +180,7 @@ impl<T: AuthenticatedMultiplexedTransport> SwarmBuilder<super::provider::Tokio, 
     > {
         self.without_quic()
             .without_any_other_transports()
-            .try_with_dns_config(cfg, opts)
+            .with_dns_config(cfg, opts)
     }
 }