+++ libp2p/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.57.0
 
-- Add `SwarmBuilder::try_with_dns_config` for fallible custom DNS resolver construction.
+- Change `SwarmBuilder::with_dns_config` to return `Result<_, libp2p_dns::ResolveError>`
+  instead of panicking when Hickory rejects resolver construction.
   See [PR 6418](https://github.com/libp2p/rust-libp2p/pull/6418).
 
 - Remove `wasm-bindgen` feature and make `wasm` support implicit.